requirements.txt
sounddevice
soundfile
numpy
openai
keyboard
python-dotenv
